+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wrapper around body of frontend.
 - Wrapper around content
 - Setting tools for global settings
+- Inject validation from the server to js
+- Inject old input to js
 
 ### Changed
 - Changed the variable names to be injected and do so through a view composer for all Bristol SU bases